+++ shared/framework-kernel/src/lib/hooks/index.ts
@@ -1,7 +1,15 @@
 import useEditFinish from './useEditFinish';
+import useMenuForQuasar from './useMenuForQuasar';
+import useMenuForVuetify from './useMenuForVuetify';
 import usePasskey from './usePasskey';
-import useSystemMenu from './useSystemMenu';
 import useSystemRoute from './useSystemRoute';
 import useSystemTheme from './useSystemTheme';
 
-export { useEditFinish, usePasskey, useSystemMenu, useSystemRoute, useSystemTheme };
+export {
+  useEditFinish,
+  usePasskey,
+  useMenuForQuasar,
+  useMenuForVuetify,
+  useSystemRoute,
+  useSystemTheme,
+};