+++ shared/framework-kernel/src/lib/hooks/useMenuForQuasar.ts
@@ -3,7 +3,7 @@ import type { RouteRecordRaw } from 'vue-router';
 import { lodash } from '@herodotus-cloud/core';
 import { useRouterStore, useAuthenticationStore } from '@herodotus-cloud/framework-kernel';
 
-export default function useSystemMenu() {
+export default function useQuasarMenu() {
   const store = useRouterStore();
   const authentication = useAuthenticationStore();