+++ shared/apis/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @herodotus-cloud/apis
 
+## 1.7.12
+
+### Patch Changes
+
+- 添加支持不同组件菜单定义
+- Updated dependencies
+  - @herodotus-cloud/core@1.7.12
+
 ## 1.7.11
 
 ### Patch Changes